requirements.txt
pytest
pytest-allure-adaptor
allure-python-commons
webium
atomicwrites
attrs
colorama
future
importlib-metadata
more-itertools
numpy
packaging
pandas
pluggy
py
pyparsing
python-dateutil
pytz
selenium
six
urllib3
waiting
wcwidth
webium
xlrd
zipp
